% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertODV.R
\name{convertODV}
\alias{convertODV}
\title{Convert a data frame to a format suitable for importing into ODV}
\usage{
convertODV(df, rep_value = -999)
}
\arguments{
\item{df}{Data frame to be converted}

\item{rep_value}{Value to replace NAs with. Defaults to -999}
}
\description{
This function replaces all NAs in a data frame with a replacement
value, by default -999
}
\examples{
convertODV(df)
}
