% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feOx.R
\name{feOx}
\alias{feOx}
\title{Iron(II) oxidation kinetics}
\usage{
feOx(temp, ph, o2, I = 0.7)
}
\arguments{
\item{temp}{Temperature in degrees C}

\item{ph}{pH}

\item{o2}{Dissolved oxygen concentration, in umol/kg}

\item{I}{Ionic strength. Defaults to 0.7 (35 psu seawater)}
}
\value{
Oxidation rate constant (kg H20 / mol Fe(II) / min)
}
\description{
Given temperature, pH, O2 concentrations, and ionic strength, calculate 
the first order reaction constant for iron(II) oxidation by oxygen. Based
on the equations of Millero et al. GCA. 1987.
}
\examples{
feOx(temp = 20, ph = 8.2, o2 = 100)
}
