% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIdp.R
\name{getIdp}
\alias{getIdp}
\title{Download the GEOTRACES 2017 Intermediate Data Product}
\usage{
getIdp()
}
\value{
the IDP, a data frame/tibble
}
\description{
The 2017 IDP, arranged in a useful format to work with in R, rather than
dealing with the files manually.
}
\examples{
idp <- getIdp()
}
