% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimKinetics.R
\name{optimKinetics}
\alias{optimKinetics}
\title{Determine an oxidation constant}
\usage{
optimKinetics(t, c)
}
\arguments{
\item{t}{time points of the experiment (vector)}

\item{c}{concentrations at the time points (vector of same size)}
}
\value{
A first order constant
}
\description{
Typically, first order reaction constants are estimated by log
transforming the data, and then using linear regression. Unfortunately,
this biases the result in favor of lower concentration samples. This
function fits a first order reaction curve to the concentration data
directly, solving this problem
}
\examples{
k <- optimKinetics(t, c)
}
