% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{Principal component analysis}
\usage{
pca(df, scale = TRUE)
}
\arguments{
\item{df}{The data frame}
}
\value{
a list containing weights of the analysis, principal components, the percent variance of each component, and the factor loadings
}
\description{
Given a data frame, perform a principal component analysis
}
