% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipTime.R
\name{shipTime}
\alias{shipTime}
\title{Time for a ship to complete its objectives}
\usage{
shipTime(lats, longs, times, speed)
}
\arguments{
\item{lats}{latitude of each station, should be from -90 to 90. Should
be vector of same size as longs}

\item{longs}{longitude of each station, should be from -180 to 180}

\item{times}{estimated time at each station in hours. Should be a vector}

\item{speed}{estimated speed of the ship, in knots}
}
\value{
estimated time in hours
}
\description{
Given a set of station coordinates, estimated time at each station,
and a speed of the ship, calculate how long it will take for the ship
to do all that
}
\examples{
shipTime(c(0, 10, 10), c(0, 10, 15), c(1, 0, 0), 12)
}
