% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potentialDensity.R
\name{sigmaTheta}
\alias{sigmaTheta}
\title{Calculate potential density}
\usage{
sigmaTheta(t, s)
}
\arguments{
\item{t}{temperature in deg C}

\item{s}{salinity, in absolute salinity units}
}
\value{
potential density in units of kg/m^3
}
\description{
Given the temperature and salinity of a seawater sample, calculate what
its potential density would be at sea level
}
\examples{
sigmaTheta(t, s)
}
