% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sectionPlot.R
\name{sectionPlot}
\alias{sectionPlot}
\title{Oceanographic section plot}
\usage{
sectionPlot(dist, depth, c, title = "", xlab = "Distance (km)",
  ylab = "Depth (m)", units = "nM", name = "Fe")
}
\arguments{
\item{dist}{distance along the section}

\item{depth}{depth of samples}

\item{c}{concentration of the element in the samples}

\item{title}{title to put on the plot, defaults to none}

\item{xlab}{label for the x axis, defaults to 'Distance (km)'}

\item{ylab}{label for the y axis, defaults to 'Depth (m)'}

\item{units}{what unit is being used for the concentraiton, defaults
to 'nM'}

\item{name}{what element is being plotted, defaults to 'Fe'}
}
\description{
Given measurements of sample locations and the concentraiton of an
element in those samples, generate a blue to red contour plot. Requires
the akima, ggplot2, and stringr packages
}
\examples{
sectionPlot(dist, depth, c)

}
